requirements.txt
Django == 1.6.5
django-rosetta == 0.7.4
django-extensions == 1.4.4
gunicorn
psycopg2
dj-static == 0.0.6
gevent-socketio
